requirement.txt
streamlit
langchain
openai
chromadb
pypdf
python-docx
sentence-transformers
pillow
openai-whisper
gtts
moviepy
opencv-python
torch
transformers
